Allow DynamicArray to accept arbitraty scalar type
Previously, DynamicArray::resize could not compile with an arbitrary
scalar type. It needs to be able to initialize the packet/array of the
scalar with zero<>() and it needs the & operator for a mask operation in
DynamicArray::clean_trailing_().

I defined a compile time check which will determine if the scalar is
supported by DynamicArray. If the check fails, the unsupported calls are
removed.

Additionally, I default the IndexScalar to uint64_t if the scalar type
is unsupported.